refactor: CCP-4535 move bceid landing page
